refactor to organize media details use cases into subpackage
