+++ domain/usecase/src/main/kotlin/usecase/mediadetails/GetMovieReviewUseCase.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.Review
 import repository.MovieDetailsRepository
