+++ domain/usecase/src/main/kotlin/usecase/mediadetails/GetSeriesCastUseCase.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.MediaCast
 import repository.TvShowDetailsRepository
