+++ domain/usecase/src/main/kotlin/usecase/mediadetails/GetSeriesReviewUseCase.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.Review
 import repository.TvShowDetailsRepository
