+++ domain/usecase/src/main/kotlin/usecase/mediadetails/GetSimilarMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.Movie
 import repository.MovieDetailsRepository
