+++ domain/usecase/src/main/kotlin/usecase/mediadetails/GetTvShowDetailsUseCase.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import repository.TvShowDetailsRepository
