+++ domain/usecase/src/test/kotlin/usecase/mediadetails/GetMovieDetailsUseCaseTest.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.MovieDetails
 import com.berlin.entity.GenreEntity
