+++ domain/usecase/src/test/kotlin/usecase/mediadetails/GetMovieGalleryUseCaseTest.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
@@ -7,7 +7,6 @@ import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
-import org.junit.jupiter.api.assertThrows
 import repository.MovieDetailsRepository
 
 class GetMovieGalleryUseCaseTest {
