+++ domain/usecase/src/test/kotlin/usecase/mediadetails/GetSeriesReviewUseCaseTest.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecase.mediadetails
 
 import com.berlin.entity.Review
 import com.google.common.truth.Truth
