+++ domain/usecase/src/test/kotlin/usecase/mediadetails/GetSimilarMoviesUseCaseTest.kt
@@ -1,25 +1,24 @@
-package usecase
+package usecase.mediadetails
 
-import com.berlin.entity.TVShow
+import com.berlin.entity.Movie
 import com.google.common.truth.Truth
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.LocalDate
-import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
-import repository.TvShowDetailsRepository
+import repository.MovieDetailsRepository
 
-class GetSimilarSeriesUseCaseTest {
+class GetSimilarMoviesUseCaseTest {
 
-    private val seriesDetailsRepository = mockk<TvShowDetailsRepository>()
-    private lateinit var getSimilarSeriesUseCase: GetSimilarSeriesUseCase
+    private val movieDetailsRepository = mockk<MovieDetailsRepository>()
+    private lateinit var getSimilarMoviesUseCase: GetSimilarMoviesUseCase
 
     @Before
     fun setUp() {
-        getSimilarSeriesUseCase = GetSimilarSeriesUseCase(seriesDetailsRepository)
+        getSimilarMoviesUseCase = GetSimilarMoviesUseCase((movieDetailsRepository))
     }
 
     @Test
@@ -27,17 +26,17 @@ class GetSimilarSeriesUseCaseTest {
         // given
         val mediaId = 0L
         coEvery {
-            seriesDetailsRepository.getSeriesSimilar(
+            movieDetailsRepository.getMovieSimilar(
                 mediaId,
             )
-        } returns getSimilarSeries()
+        } returns getSimilarMovie()
 
         //when
-        val result = getSimilarSeriesUseCase.invoke(mediaId)
+        val result = getSimilarMoviesUseCase.invoke(mediaId)
 
         // then
-        Truth.assertThat(result).isEqualTo(getSimilarSeries())
-        coVerify(exactly = 1) { seriesDetailsRepository.getSeriesSimilar(mediaId) }
+        Truth.assertThat(result).isEqualTo(getSimilarMovie())
+        coVerify(exactly = 1) { movieDetailsRepository.getMovieSimilar(mediaId) }
     }
 
     @Test
@@ -45,26 +44,27 @@ class GetSimilarSeriesUseCaseTest {
         //given
         val mediaId = 2L
         coEvery {
-            seriesDetailsRepository.getSeriesSimilar(
+            movieDetailsRepository.getMovieSimilar(
                 mediaId,
             )
         } returns emptyList()
 
         //when
-        val result = getSimilarSeriesUseCase.invoke(mediaId)
+        val result = getSimilarMoviesUseCase.invoke(mediaId)
 
         //then
         Truth.assertThat(result).isEmpty()
-        coVerify(exactly = 1) { seriesDetailsRepository.getSeriesSimilar(mediaId) }
+        coVerify(exactly = 1) { movieDetailsRepository.getMovieSimilar(mediaId) }
 
     }
 
-    private fun getSimilarSeries(): List<TVShow> {
 
-        val seriesList = mutableListOf<TVShow>()
+    private fun getSimilarMovie(): List<Movie> {
+
+        val movieList = mutableListOf<Movie>()
         for (i in 0..5) {
-            seriesList.add(
-                TVShow(
+            movieList.add(
+                Movie(
                     id = 1L,
                     title = "Inception$i",
                     rating = 8.8,
@@ -78,7 +78,7 @@ class GetSimilarSeriesUseCaseTest {
                 )
             )
         }
-        return seriesList
+        return movieList
     }
