+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mediadetails/details/MediaDetailsViewModel.kt
@@ -17,17 +17,17 @@ import com.berlin.entity.Episodes
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
-import usecase.GetMovieCastUseCase
-import usecase.GetMovieDetailsUseCase
-import usecase.GetMovieGalleryUseCase
-import usecase.GetMovieReviewUseCase
-import usecase.GetSeasonEpisodesUseCase
-import usecase.GetSeriesCastUseCase
-import usecase.GetSeriesGalleryUseCase
-import usecase.GetSeriesReviewUseCase
-import usecase.GetSimilarMoviesUseCase
-import usecase.GetSimilarSeriesUseCase
-import usecase.GetTvShowDetailsUseCase
+import usecase.mediadetails.GetMovieCastUseCase
+import usecase.mediadetails.GetMovieDetailsUseCase
+import usecase.mediadetails.GetMovieGalleryUseCase
+import usecase.mediadetails.GetMovieReviewUseCase
+import usecase.mediadetails.GetSeasonEpisodesUseCase
+import usecase.mediadetails.GetSeriesCastUseCase
+import usecase.mediadetails.GetSeriesGalleryUseCase
+import usecase.mediadetails.GetSeriesReviewUseCase
+import usecase.mediadetails.GetSimilarMoviesUseCase
+import usecase.mediadetails.GetSimilarSeriesUseCase
+import usecase.mediadetails.GetTvShowDetailsUseCase
 
 class MediaDetailsViewModel(
     savedStateHandle: SavedStateHandle,
